requirements.txt
flask==2.2.2
werkzeug==2.2.2  # or higher to match Flask 2.2.2
mysql-connector-python==8.0.30
python-dotenv==1.0.1